Changed final leaderboard to fit all players, set up framework to alow for sorting of final leaderboard by time and not order joined

+++ src/Client/ClientMaster.java
@@ -20,7 +20,7 @@ public class ClientMaster implements Runnable {
     private int obstacleCount;
     private String[] playerNames;
     private int[] obstaclesPassed;
-    private long[] playerFinishTimes;
+    private int[] playerFinishTimes;
     private int gameId;
     private boolean ready;
     private int playerId;
@@ -32,7 +32,7 @@ public class ClientMaster implements Runnable {
         commandQueue = new ArrayList<String>();
         ready = false;
         this.obstaclesPassed = new int[] {0,0};
-        this.playerFinishTimes = new long[] {0,0};
+        this.playerFinishTimes = new int[] {0,0};
         this.playerNames = new String[] {"notset", "notset2"};
         this.gameId = 0;
         this.playerId = -1;
@@ -68,7 +68,7 @@ public class ClientMaster implements Runnable {
 
                         case "getscores":
                             String[] scores = serverResponse.split("\\s");
-                            this.playerFinishTimes = new long[scores.length];
+                            this.playerFinishTimes = new int[scores.length];
                             for (int i = 0; i < scores.length; i++) {
                                 this.playerFinishTimes[i] = Integer.parseInt(scores[i]);
                             }
@@ -180,7 +180,7 @@ public class ClientMaster implements Runnable {
         commandQueue.add("getscores");
     }
 
-    public long[] getFinishTimes() {
+    public int[] getFinishTimes() {
         return this.playerFinishTimes;
     }
